split test_okvim to several files
