+++ spyder_okvim/executor/executor_sub.py
@@ -688,61 +688,3 @@ class ExecutorSearch(ExecutorSubBase):
         self.vim_status.sub_mode = None
 
         return True
-
-
-class ExecutorSubSpace(ExecutorSubBase):
-    """Submode of Space."""
-
-    def __init__(self, vim_status):
-        super().__init__(vim_status)
-        self.has_zero_cmd = False
-
-        cmds = 'itbrfp\r'
-        self.pattern_cmd = re.compile(r"(\d*)([{}])".format(cmds))
-        self.set_selection_to_editor_using_vim_selection = \
-            self.vim_status.cursor.set_selection_to_editor_using_vim_selection
-
-    def i(self, num=1, num_str=''):
-        """Insert import statements from user defined list(for personal)."""
-        editor = self.get_editor()
-        try:
-            editor.auto_import.auto_import()
-            self.vim_status.set_focus_to_vim()
-        except AttributeError:
-            pass
-
-    def b(self, num=1, num_str=""):
-        """Toggle break."""
-        self.get_editorstack().set_or_clear_breakpoint()
-
-    def enter(self, num=1, num_str=""):
-        """Run cell and advance."""
-        editor = self.get_editor()
-        editor.sig_run_cell_and_advance.emit()
-        self.vim_status.set_focus_to_vim()
-
-    def r(self, num=1, num_str=""):
-        """Run selected text or current line in console."""
-        editor = self.get_editor()
-        old_cursor = self.set_selection_to_editor_using_vim_selection()
-        editor.sig_run_selection.emit()
-        if old_cursor:
-            editor.setTextCursor(old_cursor)
-        self.vim_status.set_focus_to_vim()
-
-    def f(self, num=1, num_str=''):
-        """Format document automatically."""
-        editor = self.get_editor()
-        old_cursor = self.set_selection_to_editor_using_vim_selection()
-        try:
-            editor.format_document_or_range()
-            if old_cursor:
-                editor.setTextCursor(old_cursor)
-            self.vim_status.set_focus_to_vim()
-        except AttributeError:
-            pass
-
-    def p(self, num=1, num_str=''):
-        """Open switcher for symbol."""
-        self.vim_status.main.open_switcher()
-
